resume_builder/requirements.txt
python-dotenv
requests
python-docx
customtkinter
Pillow